mlmodels/pest_control/requirements.txt
flask==3.1.2
gunicorn==23.0.0
numpy==2.1.3
pillow==11.3.0
tensorflow==2.19.1
